refactor(community): consolidate shared widgets per review feedback
- add optional icon support to GeneralButtonV2 and drop the separate
  CommunitySubmitButton
- extract the private field label in LabeledProductTextField into a
  public CustomTextFieldLabel and drop the duplicate RequiredLabel
- rebuild CommunitySendButton on IconButton.filled
- use InkWell for tap-to-dismiss in HeroPhotoViewPage
